requirements.txt
binance-connector
requests
binance
pip==24.2
requests==2.32.3
grpcio==1.65.1
python-dotenv==1.0.1
google~=3.0.0
PyYAML>=5.4.1
redmail~=0.6.0
binance~=0.3
pip~=24.2
attrs~=23.2.0
setuptools~=65.5.0
python-binance==0.7.9
Twisted==21.2.0
Unidecode==1.2.0
NumPy==1.26.4
Flask==3.0.0
mypy==0.910
typed-ast==1.5.0
asyncio==3.4.3
grpcio-tools==1.65.1
grpcio==1.65.1
pydantic
py_directus